Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

JSOSuite.solvers

JSOSuite.select_solvers

JSOSuite.solve

​

JSOSuite.solvers — Constant

solvers

DataFrame with the solver and their properties.

For each solver, the following are available:

name::String: name of the solver;

name_solver = Symbol: name of the solver structure, :not_implemented if not implemented;

solve_function = Symbol: name of the function;

is_available = Bool: true if the solver is available as some may require a license;

bounds = Bool: true if the solver can handle bound constraints;

equalities = Bool: true if the solver can handle equality constraints;

inequalities = Bool: true if the solver can handle inequality constraints;

specialized_nls = Bool: true if the solver has a specialized variant for nonlinear least squares;

can_solve_nlp = Bool: true if the solver can solve any nlp. Some may only solve nonlinear least squares;

nonlinear_obj = Bool: true if the solver can handle nonlinear objective;

nonlinear_con = Bool: true if the solver can handle nonlinear constraints;

highest_derivative = Int: order of the highest derivative used by the algorithm.

JSOSuite.select_solvers — Function

select_solvers(nlp::AbstractNLPModel, verbose = 1, highest_derivative_available::Integer = 2)

Narrow the list of solvers to solve nlp problem using highest_derivative_available.

This function checks whether the model has:

linear or nonlinear constraints;

unconstrained, bound constraints, equality constraints, inequality constraints;

nonlinear or quadratic objective.

A linear or quadratic objective is detected if the type of nlp is a QuadraticModel or an LLSModel. The selection between a general optimization problem and a nonlinear least squares is done in solve.

If no solvers were selected, consider setting verbose to true to see what went wrong.

Output

selected_solvers::DataFrame: A subset of solvers adapted to the problem nlp.

See also solve.

Examples

using ADNLPModels, JSOSuite
nlp = ADNLPModel(x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2, [-1.2; 1.0])
selected_solvers = JSOSuite.select_solvers(nlp)
print(selected_solvers[!, :name])

JSOSuite.solve — Function

stats = solve(nlp::Union{AbstractNLPModel, JuMP.Model}; kwargs...)
stats = solve(nlp::Union{AbstractNLPModel, JuMP.Model}, solver_name::Symbol; kwargs...)

Compute a local minimum of the optimization problem nlp.

JuMP.Model are converted in NLPModels via NLPModelsJuMP.jl.

If your optimization problem has a quadratic or linear objective and linear constraints consider using QuadraticModels.jl or LLSModels.jl for the model definition.

Keyword Arguments

All the keyword arguments are passed to the selected solver. Keywords available for all the solvers are given below:

atol: absolute tolerance;

rtol: relative tolerance;

max_time: maximum number of seconds;

max_eval: maximum number of cons + obj evaluations;

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Further possible options are documented in each solver's documentation.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Examples

using ADNLPModels, JSOSuite
nlp = ADNLPModel(x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2, [-1.2; 1.0])
stats = solve(nlp, verbose = 0)
stats

The list of available solver can be obtained using JSOSuite.solvers[!, :name] or see select_solvers.

using ADNLPModels, JSOSuite
nlp = ADNLPModel(x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2, [-1.2; 1.0])
stats = solve(nlp, "DCISolver", verbose = 0)
stats